Put classes in different files, and tests as well
