+++ specs/a_spec.rb
@@ -1,0 +1,1 @@
+require_relative 'spec_helper'
