+++ specs/spec_helper.rb
@@ -1,10 +1,13 @@
+require 'simplecov'
+SimpleCov.start
+
 require 'minitest'
 require 'minitest/autorun'
 require 'minitest/reporters'
 require 'minitest/skip_dsl'
-require_relative '../lib/ride_share'
-require 'csv'
-require 'simplecov'
-SimpleCov.start
+require_relative '../lib/rider'
+require_relative '../lib/driver'
+require_relative '../lib/trip'
+
 
 Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
